+++ src/client/sdl/renderer.cpp
@@ -9,13 +9,13 @@ Renderer::Renderer(SDL2pp::Window &window) : renderer_(window,
                                                        SDL_RENDERER_ACCELERATED) {
 }
 
-void Renderer::copy(Sprite &sprite, int x, int y) {
+void Renderer::copy(Sprite &sprite, size_t x, size_t y) {
   const SDL2pp::Rect dest(x, y, sprite.width(), sprite.height());
   renderer_.Copy(sprite.sprite(), SDL2pp::NullOpt, dest);
 }
 
-void Renderer::copy(Sprite &sprite, int x_src, int y_src, int x_dst,
-                    int y_dst, int width, int height) {
+void Renderer::copy(Sprite &sprite, size_t x_src, size_t y_src, size_t x_dst,
+                    size_t y_dst, size_t width, size_t height) {
   const SDL2pp::Rect src(x_src, y_src, width, height);
   const SDL2pp::Rect dest(x_dst, y_dst, sprite.width(), sprite.height());
   renderer_.Copy(sprite.sprite(), src, dest);
@@ -31,14 +31,14 @@ SDL2pp::Renderer &Renderer::renderer() {
   return renderer_;
 }
 
-int Renderer::getMinDimension() const {
+size_t Renderer::getMinDimension() const {
   return std::min(renderer_.GetOutputWidth(), renderer_.GetOutputHeight());
 }
 
-int Renderer::getWidth() const {
+size_t Renderer::getWidth() const {
   return renderer_.GetOutputWidth();
 }
 
-int Renderer::getHeight() const {
+size_t Renderer::getHeight() const {
   return renderer_.GetOutputHeight();
 }
